Add files via upload
This code takes an input tensor, and finds the average position of the Cherenkov rings by weighing the charge logged by the PMT. It then calibrates the image such that the center of the ring is placed at the center of the image. 
Another optional function written within it would be the plotting function that outputs the image of the tensor. There is also a function that plots the projection of the image onto a 2D plane allowing for the mean to be taken.